+++ lib/shared_object.dart
@@ -1,0 +1,5 @@
+/// Shared Object Library.
+library shared_map.object;
+
+export 'src/shared_object.dart';
+export 'src/shared_reference.dart';