+++ lib/shared_object_isolate.dart
@@ -1,0 +1,5 @@
+/// Shared Object Library.
+library shared_map.object_isolate;
+
+export 'shared_object.dart';
+export 'src/shared_object_isolate.dart';